# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/languages/{languageCode}": {
      "get": {
        "summary": "Read Language By Code",
        "description": "Retrieves a language by its code.\n\nArgs:\n    language_code: The code of the language to retrieve.\n    db: Database session.",
        "operationId": "read_language_by_code_languages__languageCode__get",
        "parameters": [
          {
            "name": "languageCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Languagecode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Language": {
        "properties": {
          "languageCode": {
            "type": "string",
            "title": "Languagecode"
          },
          "languageName": {
            "type": "string",
            "title": "Languagename"
          }
        },
        "type": "object",
        "required": [
          "languageCode",
          "languageName"
        ],
        "title": "Language",
        "description": "Pydantic schema for language.\n\nAttributes:\n    language_code: String code key to language.\n    language_name: String name of language."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}